+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * Fixed client_bounding_boxes.py example script
   * Exposed in the API: camera, exposure, depth of field, tone mapper and color attributes for the RGB sensor
   * Fixed materials and semantic segmentation issues regarding importing assets
+  * Added TrafficManager to replace autopilot in managing the NPC vehicles
 
 ## CARLA 0.9.6